refactor(core): move the k8s time formatters out of the frozen design
