+++ packages/core/src/index.ts
@@ -31,6 +31,7 @@ export * from "./lib/exec";
 export * from "./lib/files";
 export * from "./lib/forward";
 export * from "./lib/helm";
+export * from "./lib/k8sTime";
 export * from "./lib/kinds";
 export * from "./lib/kubectlMapper";
 export * from "./lib/llm";
